+++ Source/GenerationsRaytracing.X86/MaterialData.h
@@ -6,7 +6,7 @@ class MaterialDataEx : public Hedgehog::Mirage::CMaterialData
 public:
     uint32_t m_materialId;
     XXH32_hash_t m_materialHash;
-    uint32_t m_hashFrame;
+    uint64_t m_hashFrame;
     boost::shared_ptr<CMaterialData> m_fhlMaterial;
 };
